# Use html_entity_decode instead of entities array