This uses the [html_entity_decode](https://www.php.net/manual/en/function.html-entity-decode.php) native function instead of loading an array of all HTML5 entities to optimize memory usage (#385).

Before the patch:
```
Benchmark Results, CPU:
        1) Parsedown                    1.24 ms              
        2) cebe/markdown extra          1.57 ms      -0.33 ms
        3) cebe/markdown                1.66 ms      -0.42 ms
        4) cebe/markdown gfm            2.05 ms      -0.81 ms
        5) PHP Markdown                 2.42 ms      -1.18 ms
        6) PHP Markdown Extra           3.51 ms      -2.27 ms
        7) CommonMark                  15.43 ms     -14.19 ms

Benchmark Results, Peak Memory:
        1) PHP Markdown                 2444 kB              
        2) Parsedown                    2587 kB       -143 kB
        3) PHP Markdown Extra           2680 kB       -236 kB
        4) cebe/markdown                2693 kB       -249 kB
        5) cebe/markdown extra          2764 kB       -320 kB
        6) cebe/markdown gfm            3118 kB       -674 kB
        7) CommonMark                   6844 kB      -4400 kB
```

After the patch:
```
Benchmark Results, CPU:
        1) Parsedown                    1.21 ms              
        2) cebe/markdown                1.35 ms      -0.14 ms
        3) cebe/markdown extra          1.52 ms      -0.31 ms
        4) cebe/markdown gfm            1.94 ms      -0.73 ms
        5) PHP Markdown                 2.35 ms      -1.14 ms
        6) PHP Markdown Extra           3.46 ms      -2.25 ms
        7) CommonMark                  15.46 ms     -14.25 ms

Benchmark Results, Peak Memory:
        1) PHP Markdown                 2444 kB              
        2) Parsedown                    2587 kB       -143 kB
        3) PHP Markdown Extra           2680 kB       -236 kB
        4) cebe/markdown                2693 kB       -249 kB
        5) cebe/markdown extra          2764 kB       -320 kB
        6) cebe/markdown gfm            3118 kB       -674 kB
        7) CommonMark                   6542 kB      -4098 kB
```

The CPU doesn't seem to be notably decreased or increased in my tests (the impact is hidden by the measurement noise) and the memory usage is decreased by ~5% (302 kB).
All the tests are passing, tested with PHP 7.3.9.