@colinodell 2019-10-13
I really like this approach!  Unfortunately, I forgot to make that array `private` before releasing 1.0.0, so we're stuck with it until 2.0.0 :-/

That being said, we can avoid that problem by creating a brand new class (named `Html5EntityDecoder`) which incorporates your changes and then deprecate the old class :wink:   I'll follow this PR up with another one which does exactly that.

Thank you so much for your contribution!